依頼のあった修正を適用
