+++ ui-test/ogp_screenshot.py
@@ -20,6 +20,7 @@ PATHS = {
     "/cards/number-of-inspection-persons": (959, 600),
     "/cards/positive-number-by-diagnosed-date":(959, 730),
     "/cards/positive-rate": (959, 730),
+    "/cards/monitoring-number-of-reports-to-covid19-consultation-desk": (959, 500),
 }
 
 options = webdriver.ChromeOptions()
